+++ transports/jetty/src/test/java/org/mule/transport/servlet/jetty/functional/JettyCustomExceptionStrategyTestCase.java
@@ -10,11 +10,9 @@ import static junit.framework.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.mule.module.http.api.client.HttpRequestOptionsBuilder.newOptions;
 
-import org.mule.DefaultMuleMessage;
 import org.mule.api.MuleMessage;
 import org.mule.api.client.MuleClient;
 import org.mule.module.http.api.client.HttpRequestOptions;
-import org.mule.module.http.api.client.HttpRequestOptionsBuilder;
 import org.mule.tck.junit4.FunctionalTestCase;
 import org.mule.tck.junit4.rule.DynamicPort;
 import org.mule.transport.http.HttpConnector;
@@ -41,6 +39,6 @@ public class JettyCustomExceptionStrategyTestCase extends FunctionalTestCase
         MuleMessage response = client.send("http://localhost:" + dynamicPort1.getNumber() + "/test", getTestMuleMessage(), httpRequestOptions);
         assertNotNull(response);
         assertNotNull(response.getInboundProperty("CustomES"));
-        assertEquals(response.getInboundProperty(HttpConnector.HTTP_STATUS_PROPERTY), 400);
+        assertEquals((int)response.getInboundProperty(HttpConnector.HTTP_STATUS_PROPERTY), 400);
     }
 }
